Split off pinExpressions and showTips
