+++ src/preload/moduleOverrides/extra-expression-buttons.replacements
@@ -35,22 +35,22 @@ Insert before the delete button but after the other buttons.
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => window.DesModder?.controller?.isPluginEnabled?.("pin-expressions") && this.model().type !== "folder"
+    predicate: () => DSM.pinExpressions && this.model().type !== "folder"
   },
   () => $DCGView.Components.IfElse(
-    () => window.DesModder?.controller?.isExpressionPinned(this.model().id),
+    () => DSM.pinExpressions?.isExpressionPinned(this.model().id),
     {
       false: DesModder?.controller?.createAction(
         "pin-expressions-pin",
         "dsm-pin-button",
         "dsm-icon-bookmark",
-        () => window.DesModder.controller.pinExpression(this.model().id)
+        () => DSM.pinExpressions?.pinExpression(this.model().id)
       ),
       true: DesModder?.controller?.createAction(
         "pin-expressions-unpin",
         "dsm-unpin-button dcg-selected",
         "dsm-icon-bookmark",
-        () => window.DesModder.controller.unpinExpression(this.model().id)
+        () => DSM.pinExpressions?.unpinExpression(this.model().id)
       )
     }
   )
@@ -59,7 +59,7 @@ $DCGView.createElement(
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => window.DesModder?.controller?.isPluginEnabled?.("folder-tools") && this.model().type === "text"
+    predicate: () => DSM.folderTools && this.model().type === "text"
   },
   DesModder?.controller?.createAction(
     "folder-tools-enclose",
@@ -72,7 +72,7 @@ $DCGView.createElement(
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => window.DesModder?.controller?.isPluginEnabled?.("folder-tools") && this.model().type === "folder"
+    predicate: () => DSM.folderTools && this.model().type === "folder"
       && window.Calc.controller.getItemModelByIndex(this.model().index + 1)?.folderId === this.model().id
   },
   DesModder?.controller?.createAction(
@@ -86,7 +86,7 @@ $DCGView.createElement(
 $DCGView.createElement(
   $DCGView.Components.If,
   {
-    predicate: () => window.DesModder?.controller?.isPluginEnabled?.("folder-tools") && this.model().type === "folder"
+    predicate: () => DSM.folderTools && this.model().type === "folder"
   },
   DesModder?.controller?.createAction(
     "folder-tools-merge",
