+++ src/preload/moduleOverrides/insert-panels.replacements
@@ -44,7 +44,7 @@ $DCGView.createElement(
     model => $DCGView.createElement(
         $DCGView.Components.If,
         {
-          predicate: () => window.DesModder?.controller?.isExpressionPinned(model.id)
+          predicate: () => DSM.pinExpressions?.isExpressionPinned(model.id)
         },
         // marking as a drag copy causes it not to affect the render shells calcuations
         // (all the logic is present already because if the top expression is dragged
